% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car2sph.R
\name{car2sph}
\alias{car2sph}
\title{Spherical coordinates}
\usage{
car2sph(X, deg = FALSE, rep = "sph", unitlength = TRUE)
}
\arguments{
\item{X}{A numerical vector or matrix whos entries constitute a cartesian coordinate.
A vector should have 3 entries and a matrix should have 3 columns.}

\item{deg}{Logical value indicating whether the output should be in randians or degrees. Radian is default.}

\item{rep}{Determine the representation of the spherical coordinates. 
If \code{"sph"} long will range from \code{0} to \code{2 * pi} and lat will range from \code{0} to \code{pi}.
The classical representation of spherical coordinates in radians.
If \code{"geo"} long will range from \code{-pi} to \code{pi} and lat will range from \code{0} to \code{pi}.
The geografical representation in radians.
Deafult is "sph".}

\item{unitlength}{Logical values indicating if the points should be returned with unit length. 
If TRUE (default) \code{r} will be set to 1.}
}
\value{
A data.frame with 3 columns longitude, latitude and radial distance.
}
\description{
Transform a 3D vector constituting a cartesian coordinate set to spherical coordinates
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
