% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_Kent_Watson_mix.R
\name{d_Kent_Watson_mix}
\alias{d_Kent_Watson_mix}
\title{Mixture Density Function of Kent and Watson}
\usage{
d_Kent_Watson_mix(x, mean_dir_Kent, minor_dir, major_dir, cons_Kent, oval_Kent,
  mean_dir_Watson, cons_Watson, p_Kent)
}
\arguments{
\item{x}{An object of class \code{\link{pps}}.}

\item{mean_dir_Kent}{Numeric vector. See \link{dKent}}

\item{minor_dir}{Numeric vector. See \link{dKent}}

\item{major_dir}{Numeric vector. See \link{dKent}}

\item{cons_Kent}{Numeric value. The consentration parameter for the Kent distribution. See \link{dKent}}

\item{oval_Kent}{Numeric value. The ovalness parameter for the Kent distribution. See \link{dKent}}

\item{mean_dir_Watson}{Numeric vector. See \link{dWatson}}

\item{cons_Watson}{Numeric value. The consentration parameter for the Watson distribution. See \link{dWatson}}

\item{p_Kent}{Numeric value between 0 and 1. The probability of observing a point from the Kent distribution.}
}
\description{
Density function for the mixture of the Kent and Watson distributions.
}
