% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dKent.R
\name{dKent}
\alias{dKent}
\title{Kent distribution density}
\usage{
dKent(x, kappa, beta, meandir = NULL, minordir = NULL, majordir = NULL)
}
\arguments{
\item{x}{A numeric vector of length 3 or a matrix with 3 columns. 
constituting the three dimensional coordinates of a point on the sphere.}

\item{kappa}{A non-negative numeric value. The consentration parameter}

\item{beta}{A numeric value. The ovalness parameter.}

\item{meandir}{A three dimensional vector. Determines where the density will be consentrated.
Default will be the northpole.}

\item{minordir}{A three dimensional vector. Determines the axis that will be the minor axis.
Default will be the y-axis.}

\item{majordir}{A three dimensional vector. Determines the axis that will be the major axis.
Default will be the x-axis.}
}
\value{
A numeric vector of length corresponding to the number of rows in x. 
The density at the points \code{X}.
}
\description{
Function for calculating the dentisty at a location x given in cartesian coordinates
}
\details{
If \code{2 * beta < kappa} then the distribution is a generalisation of the normal distribution to the sphere.
The distribution is a extension of the von-Mises Fisher distribution (\code{beta = 0}). 
\code{(meandir, minordir, majordir)} is supposed to be a orthogonal triple of unit vectors.
}
\author{
Andreas Christoffersen \email{andreas@math.aau.dk}
}
