% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.Trans.pp3.R
\name{edge.Trans.pp3}
\alias{edge.Trans.pp3}
\title{Translation Edge Correction}
\usage{
edge.Trans.pp3(X, trim = spatstat.options("maxedgewt"))
}
\arguments{
\item{X}{An object of class \code{\link{pp3}}, with a window of class \code{\link{box3}}.}

\item{trim}{option for handling the case otwo points lying on the boundary, in which case the correction factor is \code{Inf}.}
}
\value{
A numeric matrix of edge correction weights.
}
\description{
Computes Osher and Stoyans's translation edge correction weights for a 3 dimensional point pattern.
The function is only applicable for \code{\link{pp3}} objects observed on a window of class \code{\link{box3}}.
}
\details{
\code{1 / |W cup W_{eta - xi}|}
}
