% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.box3.R
\name{intersect.box3}
\alias{intersect.box3}
\title{Get intersection of two windows}
\usage{
intersect.box3(A, B, fatal = TRUE)
}
\arguments{
\item{A, B}{An object of class \code{box3}.}

\item{fatal}{Logical. Determines what happens if the intersection is empty.}
}
\value{
A window (object of class \code{\link{box3}}) or possibly \code{NULL}.
}
\description{
Function that gets the intersection of two windows of class \code{box3}.
}
