% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ksph.R
\name{Ksph}
\alias{Ksph}
\title{A modification of the Ksphere function}
\usage{
Ksph(X, r = NULL, rmax = NULL, nrval = 128, correction = "iso",
  intenss = NULL, ...)
}
\arguments{
\item{X}{An object of class \code{\link{pps}}}

\item{r}{Optional. Numeric vector of radii at which K(r) will be evaluated.}

\item{rmax}{Optional. Maximum value of argument r for which Ksph(r) will be estimated.}

\item{nrval}{Optional. Number of values of r for which Ksph(r) will be estimated.}

\item{correction}{Character string or character vector specifying the choise of edge corrections.}

\item{intenss}{Optional. A function or vector representing the intensity at the observed locations.
Only applicable for inhomgeneous point processes.}

\item{...}{Additional arguments passed to \code{intenss} if \code{intenss} is a function.}
}
\value{
An object as returned by \code{Ksphere}. An \code{\link{fv}} object.
}
\description{
Function that calculates the inhomogeneous K-functio as well as the lambdavalues required by \code{spherstat}
}
\details{
This function has been implemented since \code{\link{Ksphere}} needs an argument \code{lambdavalues} when calculating the inhomogenious K-function.
Hence, when applying the \code{\link{envelope}} function from \code{\link{spatstat}} we need to supply additional information for each point pattern for which \code{K(r)} should be calculated.
That is at this point not possible.
This function calculates the \code{lambdavalues} and the applies \code{\link{Ksphere}}.
The \code{lambdavalues} are calculated based on the function \code{FUN}.

If \code{X$window$type == "sphere"} then no edge correction is done eventhough \code{correction = "iso"}.
}
\author{
Andreas Christoffersen \email{andreas@math.aau.dk}
}
