% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pps2sp2.R
\name{pps2sp2}
\alias{pps2sp2}
\title{Convert pps object to sp2 object}
\usage{
pps2sp2(X)
}
\arguments{
\item{X}{An object of class \code{\link{pps}}.}
}
\description{
Function that converts an object of class \code{\link{pps}} to an object of class \code{\link{sp2}}.
}
