% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWatson.R
\name{rWatson}
\alias{rWatson}
\title{Simulate Watson distribution}
\usage{
rWatson(n, conc, mean)
}
\arguments{
\item{n}{Positive integer. The number of points to be simulated.}

\item{conc}{real number. The concentration parameter if negative it is a girdle distribution, if positive it is a bimodal distribution.}

\item{mean}{A real three dimensional vector constituting the mean direction of the distribution to be simulated.}
}
\value{
A matrix with the simulated data.
}
\description{
Function for simulating from a specified Watson distribution
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
