% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph2car.R
\name{sph2car}
\alias{sph2car}
\title{Cartesian coordinates}
\usage{
sph2car(X, deg = FALSE, rep = "sph")
}
\arguments{
\item{X}{A numerical vector or matrix whos entries constitute sperical coordinates.
A vector should have 2 or 3 entries and a matrix should have 2 or 3 columns.
If the input has 2 entries or columns it will be assumed that the radial distance i 1.
The order of the vector entries (matrix columns) should be longitude, latitude and radial distance (if applicable).}

\item{deg}{Logical value indicating whether the input is in randians or degrees. Radian is default.}

\item{rep}{Determine the representation of the spherical coordinates. 
If \code{"sph"} long will range from \code{0} to \code{2 * pi} and lat will range from \code{0} to \code{pi}.
The classical representation of spherical coordinates in radians.
If \code{"geo"} long will range from \code{-pi} to \code{pi} and lat will range from \code{0} to \code{pi}.
The geografical representation in radians.
Deafult is "sph".}
}
\value{
A \code{data.frame} with 3 columns representing the cartesian coordinate set.
}
\description{
Transform a spherical coordinates set to cartesian coordinates
}
\author{
Andreas Dyreborg Christoffersen \email{andreas@math.aau.dk}
}
