% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.box3.R
\name{translate.box3}
\alias{translate.box3}
\title{Translate window}
\usage{
translate.box3(x, win)
}
\arguments{
\item{x}{A vector with 3 entries. must be given in the order \code{x}, \code{y}, and \code{z}.}

\item{win}{An object of class \code{box3}, that is to be translated.}
}
\value{
An object of class \code{box3}.
}
\description{
Function that translate a window of class \code{box3} by some specified vector \code{x}.
}
