% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWatson.R
\name{dWatson}
\alias{dWatson}
\title{Watson distribution density}
\usage{
dWatson(x, kappa, alpha = NULL)
}
\arguments{
\item{x}{A numeric vector of length 3 or a matrix with 3 columns. constituting the three dimensional coordinates of a point on the sphere.}

\item{kappa}{A numeric value. The consentration parameter. May be negative.}

\item{alpha}{A numeric vector of lengthe 3. Determines the axis at which there will be rotationally symmetry.}
}
\value{
A numeric vector of length corresponding to the number of rows in x. 
The density at the points \code{X}.
}
\description{
Function for calculating the dentisty at a location x given in cartesian coordinates.
}
\details{
If \code{kappa < 0} the distribution will be an equatorial girdle.
If \code{kappa > 0} the distribution will be bimodal around the poles \code{alpha} and \code{-alpha}.
}
\author{
Andreas Christoffersen \email{andreas@math.aau.dk}
}
